Extract KernelHandler classes to separate files.
