+++ qmeq/approach/base/c_lindblad.pyx
@@ -30,7 +30,7 @@ from libc.math cimport sqrt
 from ...specfunc.c_specfunc cimport func_pauli
 
 from ..c_aprclass cimport Approach
-from ..c_aprclass cimport KernelHandler
+from ..c_kernel_handler cimport KernelHandler
 
 # ---------------------------------------------------------------------------------------------------
 # Lindblad approach
